requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT